Create day7 (UNIX)
PART 1:
"To begin, find all of the directories with a total size of at most 100000, then calculate the sum of their total sizes. In the example above, these directories are a and e; the sum of their total sizes is 95437 (94853 + 584). (As in this example, this process can count files more than once!)
Find all of the directories with a total size of at most 100000. What is the sum of the total sizes of those directories?"

PART 2:
"Find the smallest directory that, if deleted, would free up enough space on the filesystem to run the update. What is the total size of that directory?"

https://adventofcode.com/2022/day/7